Implemented adding new recipes

+++ backend/controllers/userController.js
@@ -74,6 +74,10 @@ export const deleteUser = async (req, res) => {
   const { id } = req.params;
   const tokenUserId = req.userId;
 
+  if (id !== tokenUserId) {
+    return res.status(403).json({ error: "Not Authorized!" });
+  }
+
   try {
     await prisma.user.delete({
       where: { id },
